Create useDeleteRoom custom hook
