+++ src/components/elements/deleteRoomButton/DeleteRoomButton.tsx
@@ -1,21 +1,13 @@
 import React from "react";
-import { getFirestore, doc, deleteDoc } from "firebase/firestore";
-import { app } from "../../../firebase/firebase";
-
+import { useDeleteRoom } from "../../../hooks/useDeleteRoom";
 import { ChatRoomProps } from "../../../typings/ChatRoomProps";
 
 export const DeleteRoomButton: React.FC<ChatRoomProps> = ({ roomId }) => {
-  const firestore = getFirestore(app);
+  const { deleteRoom, loading } = useDeleteRoom();
 
-  const deleteRoom = async () => {
-    try {
-      await deleteDoc(doc(firestore, "rooms", roomId));
-      alert("Room deleted successfully");
-    } catch (error) {
-      console.error("Error deleting room: ", error);
-      alert("Failed to delete room");
-    }
-  };
-
-  return <button onClick={deleteRoom}>Delete Room</button>;
+  return (
+    <button onClick={() => deleteRoom(roomId)} disabled={loading}>
+      {loading ? "Deleting..." : "Delete Room"}
+    </button>
+  );
 };
